+++ plugins/ucc_chat/lib/ucc_chat_web/channels/user_channel/side_nav/directs.ex
@@ -17,18 +17,18 @@ defmodule UccChatWeb.UserChannel.SideNav.Directs do
 
   def open_direct_channel(socket, username) do
     assigns = socket.assigns
-    with user when not is_nil(user) <- UccChat.ServiceHelpers.get_user_by_name(username),
+    with friend when not is_nil(friend) <- UccChat.ServiceHelpers.get_user_by_name(username),
          user_id <- socket.assigns.user_id,
-         false <- user.id == user_id do
+         false <- friend.id == user_id do
       direct =
-        case get_direct(user_id, username) do
+        case get_direct(user_id, friend.id) do
           nil ->
-            ChannelService.add_direct(username, user_id, nil)
-            get_direct(user_id, username)
+            ChannelService.add_direct(friend, user_id, nil)
+            get_direct(user_id, friend.id)
           direct ->
             direct
         end
-      Channels.open_room socket, assigns.room, direct.channel.name, direct.users
+      Channels.open_room socket, assigns.room, direct.channel.name, friend.username
     else
       _ ->
         Client.toastr socket, :error, ~g(Could not open that direct channel)
@@ -36,7 +36,7 @@ defmodule UccChatWeb.UserChannel.SideNav.Directs do
     socket
   end
 
-  defp get_direct(user_id, name) do
-    Direct.get_by user_id: user_id, users: name, preload: [:channel]
+  defp get_direct(user_id, friend_id) do
+    Direct.get_by user_id: user_id, friend_id: friend_id, preload: [:channel]
   end
 end