+++ plugins/ucc_chat/lib/ucc_chat_web/flex_bar/tab/user_info.ex
@@ -25,12 +25,12 @@ defmodule UccChatWeb.FlexBar.Tab.UserInfo do
   def args(socket, {user_id, channel_id, _, _}, _) do
     current_user = Helpers.get_user! user_id
     channel = Channel.get!(channel_id)
-    direct = Direct.get_by user_id: user_id, channel_id: channel_id
+    direct = Direct.get_by user_id: user_id, channel_id: channel_id, preload: [friend: [:account, :roles, user_roles: [:role]]]
 
-    user = Helpers.get_user_by_name(direct.users)
+    # user = Helpers.get_user_by_name(direct.users)
     user_info = user_info(channel, direct: true)
     {[
-      user: user,
+      user: direct.friend,
       current_user: current_user,
       channel_id: channel_id,
       user_info: user_info